updated the function, looks pretty complicated but works for the most part. Issues:
How to work when the MC  simulation outputs 2 values (f.e. OLS and GLS)? purrr package / map function seems to have issues with that. For now I changed the MC_simulation to output just one value, namely the difference between OLS and GLS variance.

Another problem is how to combine functionals with a list of functions